pyproject.toml
[project]
name = "fuiou-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "oss2>=2.19.1",
    "pycryptodome>=3.23.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
